VERY LIGHT CONFIG CHANGE

+++ include/bitvector.h
@@ -15,7 +15,7 @@ typedef uint32_t BVCTYPE;
 
 #ifdef VERY_LIGHT_FIELDS
 typedef uint16_t SZ_BITS_T;
-typedef uint8_t ALLOC_TAG_T;
+typedef uint16_t ALLOC_TAG_T;
 typedef uint8_t CONTAINER_SZ_T;
 #elif defined(LIGHT_FIELDS)
 typedef uint32_t SZ_BITS_T;
